gnmi/test_gnoi_killprocess.py add swss status check and ignore some logs
`swss` takes a longer time to restart, but the test
does not ensure the swss service is ready before proceeding to the next
test. This causes flaky errors to occur in the `pmon` test case, which
is run after `swss`.

Restarting `swss` (and some other services) also causes a number of
syslog errors (similar to how it is caused by config_reload). This is
expected, extending log_ignore logre to allow this error log.